multiple bugs in update-monitors
1) the detect clone check triggered when the two screens had the same
height.  a non-cloned screen does not necessarily (actually usually does
not with my preferences) add to the overall screen size. (I always put
outputs next to each other horizontally, not vertically)

2) the "primary" label may be used on a non-participating output.
we pick the first participating output as "primary" instead.

3) since the output to add may be listed before we know what the primary
is, we need to wait to do the add until we have read the full xrands
output.  this means we will never add more than one output per run,
but I don't think that is a big issue...